feat(portfolio): details column
