+++ packages/lib/modules/pool/PoolList/PoolListTable/PoolListTableRow.tsx
@@ -1,6 +1,6 @@
 import { Box, Grid, GridItem, GridProps, HStack, Text } from '@chakra-ui/react'
 import Link from 'next/link'
-import { getPoolPath, getPoolTypeLabel } from '../../pool.utils'
+import { getPoolPath } from '../../pool.utils'
 import MainAprTooltip from '@repo/lib/shared/components/tooltips/apr-tooltip/MainAprTooltip'
 import { memo } from 'react'
 import { NetworkIcon } from '@repo/lib/shared/components/icons/NetworkIcon'
@@ -10,11 +10,8 @@ import { PoolListTokenPills } from '../PoolListTokenPills'
 import { getUserTotalBalanceUsd } from '../../user-balance.helpers'
 import FadeInOnView from '@repo/lib/shared/components/containers/FadeInOnView'
 import { usePoolList } from '../PoolListProvider'
-import { PoolVersionTag } from './PoolVersionTag'
-import { isBoosted } from '../../pool.helpers'
 import { TokenIcon } from '@repo/lib/modules/tokens/TokenIcon'
-import { useErc4626Metadata } from '../../../erc4626/Erc4626MetadataProvider'
-import Image from 'next/image'
+import { PollListTableDetailsCell } from '@repo/lib/modules/pool/PoolList/PoolListTable/PollListTableDetailsCell'
 
 interface Props extends GridProps {
   pool: PoolListItem
@@ -54,9 +51,6 @@ export function PoolListTableRow({ pool, keyValue, ...rest }: Props) {
   } = usePoolList()
 
   const { toCurrency } = useCurrency()
-  const { getErc4626Metadata } = useErc4626Metadata()
-
-  const erc4626Metadata = getErc4626Metadata(pool)
 
   return (
     <FadeInOnView>
@@ -88,21 +82,7 @@ export function PoolListTableRow({ pool, keyValue, ...rest }: Props) {
               {displayType === PoolListDisplayType.Name && <PoolName pool={pool} />}
             </GridItem>
             <GridItem minW="32">
-              <HStack>
-                <PoolVersionTag pool={pool} />
-                <Text fontWeight="medium" textAlign="left" textTransform="capitalize">
-                  {isBoosted(pool) ? 'Boosted' : getPoolTypeLabel(pool.type)}
-                </Text>
-                {erc4626Metadata.map(metadata => (
-                  <Image
-                    alt={metadata.name}
-                    height={20}
-                    key={metadata.name}
-                    src={metadata.iconUrl || ''}
-                    width={20}
-                  />
-                ))}
-              </HStack>
+              <PollListTableDetailsCell pool={pool} />
             </GridItem>
             {userAddress ? (
               <GridItem>
