+++ packages/lib/modules/portfolio/PortfolioTable/PortfolioTable.tsx
@@ -49,7 +49,7 @@ export const portfolioOrderBy: {
 
 const rowProps = {
   px: [0, 4],
-  gridTemplateColumns: `32px minmax(320px, 1fr) repeat(2, 100px) 120px 130px 170px`,
+  gridTemplateColumns: `32px minmax(320px, 1fr) repeat(2, 120px) 120px 130px 170px`,
   alignItems: 'center',
   gap: { base: 'xxs', xl: 'lg' },
 }
